pyproject.toml
[project]
name = "mysql-library-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysql-connector-python>=9.2.0",
    "pandas>=2.2.3",
    "pandas-stubs>=2.2.3.250308",
    "plotly>=6.0.0",
    "streamlit>=1.43.2",
    "streamlit-extras>=0.6.0",
]
